Issue #17 Moved io from under utils to first class and renamed as inputoutput as it conflicts with io package in python
