+++ code/inputoutput/test_FileReader.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 import mock
-from utils.io.FileReader import FileReader
+from inputoutput.FileReader import FileReader
 
 
 class TestFileReader(TestCase):
