+++ code/utils/helpers/metahelper.py
@@ -2,7 +2,7 @@ import logging as LOG
 
 from config.AppConfiguration import APP_LOGO_DAT
 from utils.encoding.Base64 import decodeString
-from utils.io.FileReader import FileReader
+from inputoutput import FileReader
 
 
 def printLogo() -> None:
